product cart completed

+++ src/main/java/com/heroku/spacey/services/ProductService.java
@@ -19,5 +19,5 @@ public interface ProductService {
 
     void cancelProduct(Long id);
 
-    boolean checkAmount(Long productId, Long sizeId, double amount);
+    boolean checkAmount(Long productId, Long sizeId, int amount);
 }
